refactor(process): migrate connectionscheck component to v2
