+++ comp/process/bundle.go
@@ -15,7 +15,7 @@ import (
 	connectionsforwarderfx "github.com/DataDog/datadog-agent/comp/forwarder/connectionsforwarder/fx"
 	"github.com/DataDog/datadog-agent/comp/process/agent/agentimpl"
 	"github.com/DataDog/datadog-agent/comp/process/apiserver"
-	"github.com/DataDog/datadog-agent/comp/process/connectionscheck/connectionscheckimpl"
+	connectionscheckfx "github.com/DataDog/datadog-agent/comp/process/connectionscheck/fx"
 	"github.com/DataDog/datadog-agent/comp/process/containercheck/containercheckimpl"
 	expvarsfx "github.com/DataDog/datadog-agent/comp/process/expvars/fx"
 	forwardersfx "github.com/DataDog/datadog-agent/comp/process/forwarders/fx"
@@ -43,7 +43,7 @@ func Bundle() fxutil.BundleOptions {
 		profilerimpl.Module(),
 
 		// Checks
-		connectionscheckimpl.Module(),
+		connectionscheckfx.Module(),
 		containercheckimpl.Module(),
 		processcheckfx.Module(),
 		rtcontainercheckimpl.Module(),
