+++ comp/process/connectionscheck/def/component.go
